backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.3.0
PyPDF2==3.0.1
pydantic==2.7.1
groq==0.9.0
gTTS==2.5.1
kokoro-onnx>=0.4.0
soundfile>=0.12.1
faster-whisper>=1.0.3
# TTS>=0.22.0
